requirements.txt
# Core dependencies
langchain==0.3.12
langchain-community==0.3.12
langchain-openai==0.2.10

# Document processing
pypdf==5.1.0
pdfplumber==0.11.4
PyPDF2==3.0.1

# Embeddings and Vector DB
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
chromadb==0.5.23

# Web framework
fastapi==0.115.6
uvicorn==0.34.0
streamlit==1.41.0

# Data processing
numpy==1.26.4
pandas==2.2.3
python-multipart==0.0.20

# Environment and utilities
python-dotenv==1.0.1
pydantic==2.10.4
typing-extensions==4.12.2

# Korean NLP (optional)
kiwipiepy==0.20.3
konlpy==0.6.0